Extract ModelCardPanel to it's own header
